refactor(harnesstui): extract tool and status foundations
