+++ src/loushang/harnesstui/status/__init__.py
@@ -1,0 +1,1 @@
+"""Shared terminal status interaction."""
